+++ R/keep.labels.R
@@ -1,11 +1,15 @@
 
 #' Keep Labels
 #'
-#' Keep the \code{'label'} attribute on an R object when subsetting.
+#' Keep the \code{'label'} attribute on an R object when subsetting. \code{loosen.labels} allows the \code{'label'}
+#'   attribute to be lost again.
 #'
 #' @param x An R object
+#' @param i,value See \code{\link{[<-}}.
 #' @param ... Other arguments (not in use at this time).
-#' @return A copy of \code{x} with a "keep labels" class appended on.
+#' @return A copy of \code{x} with a "keep_labels" class appended on or removed. Note that for the \code{data.frame} method,
+#'   only classes on the columns are changed; the \code{data.frame} won't have an extra class appended. This is different from previous
+#'   versions of \code{arsenal}.
 #' @author Ethan Heinzen
 #' @seealso \code{\link{labels}}
 #' @name keep.labels
@@ -24,7 +28,6 @@ keep.labels <- function(x, ...)
 keep.labels.data.frame <- function(x, ...)
 {
   x[] <- lapply(x, keep.labels)
-  class(x) <- c("keep_labels_df", class(x)[class(x) != "keep_labels_df"])
   x
 }
 
@@ -42,4 +45,36 @@ keep.labels.default <- function(x, ...)
 {
   y <- NextMethod()
   keep.labels(set_attr(y, "label", attr(x, "label")))
+}
+
+#' @rdname keep.labels
+#' @export
+`[<-.keep_labels` <- function(x, i, value)
+{
+  x <- loosen.labels(x)
+  out <- NextMethod()
+  keep.labels(out)
+}
+
+#' @rdname keep.labels
+#' @export
+loosen.labels <- function(x, ...)
+{
+  UseMethod("loosen.labels")
+}
+
+#' @rdname keep.labels
+#' @export
+loosen.labels.data.frame <- function(x, ...)
+{
+  x[] <- lapply(x, loosen.labels)
+  x
+}
+
+#' @rdname keep.labels
+#' @export
+loosen.labels.default <- function(x, ...)
+{
+  class(x) <- class(x)[class(x) != "keep_labels"]
+  x
 }