+++ NEWS.md
@@ -17,6 +17,19 @@
 
 * `comparedf.control()`: allow for named `tol.vars=` argument to manually match column names. (#165)
 
+* `freqlist()`: fixed a bug where labels would get dropped when using the `subset=` argument. (#184)
+
+* Added a new function `loosen.labels()` which removes the classes added by `keep.labels()` and thereby speeds up subsetting when
+  labels are no longer needed. This is now used in `tableby()`, `modelsum()`, `freqlist()`, and `paired()`.
+  
+* `keep.labels()` no longer sticks another class on data.frames.
+
+* Fixed a bug with replacement for objects of class `"keep_labels"`.
+  
+* `tableby()` / `paired()`: fixed a bug with per-variable stats and digit specifications being lost when using the `subset=` argument. (#182, #183)
+
+* `tableby()` / `paired()`: Made all-NA summaries prettier. (#190)
+
 # arsenal v2.0.0
 
 There is a new class system (`"arsenal_table"`) which unifies `tableby()`, `modelsum()`, and `freqlist()`.